Makefile(lint): replace black with ruff
`black` is a code formatter, it only deals with stylistic problems.
So instead of `black`, use a proper linter to identify problems with the
code.

The choice here was `ruff`, which is orders of magnitude faster than
other tools, while including most rules of those tools if configured as
such.

Also adds `.ruff_cache` to `.gitignore` and `make clean`.